Move windowsStoreLocator.ts
